update process training data

+++ data_loader1.py
@@ -45,9 +45,9 @@ def prepare_training_data():
             conversation_words = re.findall(word_regex, conversation)
             # conversation: fil id of the question word (not the word), same length with max conversation length
             # keep zero at the emd if length less than max
-            conversation_ids = np.zeros(max_conversation_length)
+            conversation_ids = np.zeros(max_conversation_length, dtype=np.int8)
             for index, word in enumerate(conversation_words):
-                conversation_ids[index] = int(conversation_vocab[word])
+                conversation_ids[index] = conversation_vocab[word]
 
             training_data.append({
                 'image_id': image_id,
